CR changes: 1. move all types to comrv_types.h 2. move all #defines to comrv_defines.h 3. in comrvStackFrame_t add one padding of u16_t
